Update workflow schedule to update feeds every 10 minutes
I've noticed that more and more users are forking this repository just to reduce the time it takes to update the feeds. I don't mind this, but since this is a public repository and we don't really need to save on action usage, we might as well do it ourselves so that people don't feel like they have to create a fork for such a minor change.